requirements.txt
beautifulsoup4==4.4.1
bs4==0.0.1
lxml==3.6.0
nltk==3.2.1
oauthlib==1.0.3
Pillow==3.2.0
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
tweepy==3.5.0
